refactor: 557 - extract info-cell in separate component
